requirements.txt
numpy
pandas
torch
torchvision
PyYAML
matplotlib
onnx # Optional
onnxruntime # Optional
tqdm
pillow
scikit-learn
albumentations